Create the header on the landing page
Create the index.html and style.css files inside of TOP-Project-Landing
-Page repository. Create images repository.
Create the header of the website and the main content of the website.
The header and website are inside of their own divs and place both of
the divs inside the bigger div, so that you can make that div a flexbox
parent and nest the two children inside of it. You will have a column
flexbox, whose second flex element will be a horizontal flexbox containing
two horizontal elements.