Add the footer and a quote section to the website.
Add a margin to the link list, so that the link list is on the right
position in order to improve the visual appeal of the website.
Add a line break and the request to sign up for the newsletter in
order to meet the requirements of the course.
Add the margin to the line breaks in order to space out the website
elements and produce a more appealing layout.
Change the link list elements namings to the realistic ones.
Change the font size of the image descriptions to improve the aesthetics
of the webpage.